[F] refactor helpers
