+++ packages/frontend/src/presenter/DataSources/helpers/general.ts
@@ -1,3 +1,5 @@
-import { getProp } from '../../../utils';
+import { getProp, compose } from '../../../utils';
 
 export const getProperties = getProp('properties');
+export const getHandler = compose(getProperties, getProp('handler'));
+export const getTransforms = compose(getProperties, getProp('items'), getProp('transforms'));
