+++ packages/frontend/src/presenter/DataSources/helpers/stringsHas.ts
@@ -1,0 +1,4 @@
+import { compose, hasIndex, toLowerCase} from '../../../utils';
+
+export const stringHas = (query: string) =>
+  compose(hasIndex(toLowerCase(query)), toLowerCase)
