+++ packages/frontend/src/utils/hasIndex.ts
@@ -1,0 +1,5 @@
+type Entity = {
+  indexOf(item: any): number;
+};
+
+export const hasIndex = (value: any) => (item: Entity) => item.indexOf(value) !== -1;
