+++ packages/frontend/src/utils/toLowerCase.ts
@@ -1,0 +1,1 @@
+export const toLowerCase = (str: string) => str.toLocaleLowerCase();
